# public/openapi.yaml
openapi: 3.0.3
info:
  title: Lumigift API
  description: |
    REST API for Lumigift — time-locked cash gifts on the Stellar blockchain.

    **Authentication:** Most endpoints require a valid NextAuth.js session cookie.
    Obtain a session by completing the phone OTP flow (`POST /api/v1/auth/send-otp`
    followed by NextAuth credentials sign-in).

    **Base URL:** `https://www.lumigift.com`
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://www.lumigift.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/gifts:
    get:
      tags: [Gifts]
      summary: List gifts sent by the authenticated user
      description: Returns a cursor-paginated list of gifts (10 per page).
      security:
        - sessionCookie: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            description: Opaque pagination cursor from the previous response.
      responses:
        "200":
          description: Paginated gift list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          gifts:
                            type: array
                            items:
                              $ref: "#/components/schemas/Gift"
                          nextCursor:
                            type: string
                            nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Gifts]
      summary: Create a new gift
      description: |
        Creates a gift record and returns a Paystack payment URL.
        The gift status is `pending_payment` until the webhook confirms payment.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiftRequest"
      responses:
        "201":
          description: Gift created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          gift:
                            $ref: "#/components/schemas/Gift"
                          paymentUrl:
                            type: string
                            format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/gifts/{id}/claim:
    post:
      tags: [Gifts]
      summary: Claim a gift
      description: |
        Transfers the locked USDC to the recipient's Stellar address.
        Requires the gift to be in `locked` status and past its unlock time.
      parameters:
        - $ref: "#/components/parameters/GiftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [giftId, recipientStellarKey]
              properties:
                giftId:
                  type: string
                  format: uuid
                recipientStellarKey:
                  type: string
                  description: Recipient's Stellar public key (G...)
                  example: "GABC...XYZ"
      responses:
        "200":
          description: Gift claimed — USDC transferred
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          txHash:
                            type: string
                            description: Stellar transaction hash
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/payments/stripe/webhook:
    post:
      tags: [Payments]
      summary: Stripe webhook
      description: |
        Receives `payment_intent.succeeded` events from Stripe.
        Signature verified via `stripe-signature` header.
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Event acknowledged
        "400":
          description: Invalid signature or payload
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie obtained after OTP login.
    cronBearer:
      type: http
      scheme: bearer
      description: Static secret set in `CRON_SECRET` environment variable.
...
  parameters:
    GiftId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Gift UUID
  schemas:
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies by endpoint)
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Unauthorized"
    Gift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
        recipientName:
          type: string
        recipientPhone:
          type: string
        amountNgn:
          type: number
        message:
          type: string
          nullable: true
        mediaUrl:
          type: string
          format: uri
          nullable: true
        unlockAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending_payment, locked, claimed, cancelled, expired]
        createdAt:
          type: string
          format: date-time
...
    CreateGiftRequest:
      type: object
      required: [recipientName, recipientPhone, amountNgn, unlockAt]
      properties:
        recipientName:
          type: string
          example: "Amaka Obi"
        recipientPhone:
          type: string
          example: "+2348012345678"
        amountNgn:
          type: number
          minimum: 100
          example: 5000
        message:
          type: string
          maxLength: 500
          nullable: true
        mediaUrl:
          type: string
          format: uri
          nullable: true
        unlockAt:
          type: string
          format: date-time
          description: ISO 8601 datetime when the gift unlocks
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...